+++ domain/src/main/kotlin/omok/Application.kt
@@ -1,0 +1,9 @@
+package omok
+
+fun main() {
+    runCatching {
+        Controller(GameView(), OmokGame()).gameStart()
+    }.onFailure { exception ->
+        GameView().printErrorMessage(exception)
+    }
+}
