+++ domain/src/main/kotlin/omok/Hand.kt
@@ -1,0 +1,11 @@
+package omok
+
+class Hand {
+    private val _stones = mutableListOf<Stone>()
+    val stones
+        get() = _stones.toList()
+
+    fun add(stone: Stone) {
+        _stones.add(stone)
+    }
+}
