+++ domain/src/main/kotlin/omok/Player.kt
@@ -1,0 +1,8 @@
+package omok
+
+class Player(val hand: Hand = Hand()) {
+
+    fun put(stone: Stone) {
+        hand.add(stone)
+    }
+}
