+++ domain/src/main/kotlin/omok/Stone.kt
@@ -1,0 +1,5 @@
+package omok
+
+class Stone(val position: Position) {
+    fun findPosition(value: Position) = (position == value)
+}
