+++ domain/src/main/kotlin/omok/state/State.kt
@@ -1,0 +1,3 @@
+package omok.state
+
+sealed interface State
