+++ domain/src/main/kotlin/omok/state/Turn.kt
@@ -1,0 +1,7 @@
+package omok.state
+
+sealed class Turn : State {
+
+    object Black : Turn()
+    object White : Turn()
+}
