+++ domain/src/main/kotlin/omok/state/Win.kt
@@ -1,0 +1,7 @@
+package omok.state
+
+sealed class Win : State {
+
+    object Black : Win()
+    object White : Win()
+}
