+++ domain/src/test/kotlin/omok/PositionTest.kt
@@ -1,0 +1,18 @@
+package omok
+
+import org.assertj.core.api.AssertionsForClassTypes.assertThat
+import org.junit.jupiter.api.Test
+
+class PositionTest {
+    @Test
+    fun `해당 위치에 돌을 둘 수 있는지 판단한다`() {
+        // given
+        val position = Position(HorizontalAxis.H, 1)
+
+        // when
+        val placeable = position.isEmpty()
+
+        // then
+        assertThat(placeable).isTrue
+    }
+}
