--- src/main/kotlin/omok/domain/Hand.kt
@@ -1,11 +1,0 @@
-package omok.domain
-
-class Hand {
-    private val _stones = mutableListOf<Stone>()
-    val stones
-        get() = _stones.toList()
-
-    fun add(stone: Stone) {
-        _stones.add(stone)
-    }
-}
