--- src/main/kotlin/omok/domain/Player.kt
@@ -1,8 +1,0 @@
-package omok.domain
-
-class Player(val hand: Hand = Hand()) {
-
-    fun put(stone: Stone) {
-        hand.add(stone)
-    }
-}
