--- src/main/kotlin/omok/domain/Stone.kt
@@ -1,5 +1,0 @@
-package omok.domain
-
-class Stone(val position: Position) {
-    fun findPosition(value: Position) = (position == value)
-}
