--- src/main/kotlin/omok/domain/state/State.kt
@@ -1,3 +1,0 @@
-package omok.domain.state
-
-sealed interface State
