--- src/main/kotlin/omok/domain/state/Turn.kt
@@ -1,7 +1,0 @@
-package omok.domain.state
-
-sealed class Turn : State {
-
-    object Black : Turn()
-    object White : Turn()
-}
