--- src/main/kotlin/omok/domain/state/Win.kt
@@ -1,7 +1,0 @@
-package omok.domain.state
-
-sealed class Win : State {
-
-    object Black : Win()
-    object White : Win()
-}
